+++ engine/src/main/java/de/instinct/engine/fleet/data/StaticData.java
@@ -7,7 +7,7 @@ public class StaticData {
 
 	public long maxGameTimeMS;
 	public float zoomFactor;
-	public double atpToWin;
+	public float atpToWin;
 	public float ancientPlanetResourceDegradationFactor;
 	
 }
