+++ engine/src/main/java/de/instinct/engine/fleet/entity/projectile/ShieldDamageInstance.java
@@ -6,6 +6,6 @@ import lombok.Data;
 public class ShieldDamageInstance {
 
 	private int shieldId;
-	private double damage;
+	private float damage;
 	
 }
