+++ engine/src/main/java/de/instinct/engine/fleet/entity/unit/component/data/ShieldData.java
@@ -7,7 +7,7 @@ import lombok.ToString;
 public class ShieldData extends ComponentData {
 	
 	public ShieldType type;
-	public double strength;
-	public double generation;
+	public float strength;
+	public float generation;
 
 }
