+++ engine/src/main/java/de/instinct/engine/fleet/stats/model/unit/component/types/CoreStatistic.java
@@ -12,6 +12,6 @@ public class CoreStatistic extends ComponentStatistic {
 
 	private int timesBuilt;
 	private int timesDestroyed;
-	private double resourcesUsed;
+	private float resourcesUsed;
 
 }
